🎨 style: Lintエラーを修正
- リファクタリング済みManagerファイルを元のファイルに統合
- 重複ファイル（-refactored.ts）を削除
- mod修飾キーの判定ロジックを改善
- スペースキーのイベント発火を追加
- すべてのテストが成功することを確認
