+++ packages/input-manager/src/keyboard-manager.ts
@@ -1,94 +1,53 @@
-import type {
-	CommandHandler,
-	CommandRegistry,
-	IKeyboardManager,
-	KeyBinding,
-	KeyBindings,
-	KeyboardConfig,
-	KeyboardContext,
-	KeyboardPreset,
-} from "./types";
-import { formatShortcutLabel, ScreenReaderAnnouncer } from "./utils/accessibility";
+import { BaseInputManager, ContextMixin } from "./base-manager";
+import type { KeyBindings, KeyboardConfig, KeyboardPreset } from "./types";
+import type { IUnifiedKeyboardManager } from "./types/unified-types";
+import { ScreenReaderAnnouncer } from "./utils/accessibility";
 
-export class KeyboardManager implements IKeyboardManager {
-	private bindings: Map<string, KeyBinding>;
-	private contexts: Map<string, KeyboardContext>;
-	private activeContext: string = "default";
-	private commandHandlers: CommandRegistry;
-	private config: KeyboardConfig;
-	private listeners: Map<string, Set<(data: any) => void>>;
-	private pressedKeys: Set<string>;
-	private screenReader: ScreenReaderAnnouncer | null = null;
+/**
+ * KeyboardManager
+ * BaseInputManagerとContextMixinを使用した実装
+ */
+export class KeyboardManager
+	extends BaseInputManager<KeyboardConfig, string[], KeyBindings, KeyboardPreset, KeyboardEvent>
+	implements IUnifiedKeyboardManager
+{
+	private contextMixin = new ContextMixin();
+	private modifierState = new Map<string, boolean>();
+	private announcer?: ScreenReaderAnnouncer;
+	private commandCache = new WeakMap<KeyboardEvent, Set<string>>();
 
-	constructor(config?: KeyboardConfig) {
-		this.bindings = new Map();
-		this.contexts = new Map();
-		this.commandHandlers = {};
-		this.listeners = new Map();
-		this.pressedKeys = new Set();
-		this.config = {
-			enableInInput: false,
-			debug: false,
-			...config,
-		};
-
+	constructor(config?: Partial<KeyboardConfig>) {
+		super(config);
 		if (config?.preset) {
 			this.loadPreset(config.preset);
 		}
-
-		// デフォルトコンテキストを作成
-		this.contexts.set("default", {
-			name: "default",
-			priority: 0,
-		});
-
-		// アクセシビリティ機能を有効化
-		if (typeof window !== "undefined" && config?.enableAccessibility !== false) {
-			this.screenReader = new ScreenReaderAnnouncer();
+		if (config?.enableAccessibility) {
+			this.announcer = new ScreenReaderAnnouncer();
 		}
+		this.initialize(this.config);
+	}
+
+	getDefaultConfig(): KeyboardConfig {
+		return {
+			debug: false,
+			enableAccessibility: false,
+		};
 	}
 
 	initialize(config: KeyboardConfig): void {
 		this.config = { ...this.config, ...config };
-
 		if (config.preset) {
 			this.loadPreset(config.preset);
 		}
 		if (config.customBindings) {
-			this.setBindings(config.customBindings);
+			Object.entries(config.customBindings).forEach(([command, keys]) => {
+				this.setBinding(command, keys);
+			});
 		}
-	}
-
-	setBinding(command: string, keys: string[]): void {
-		this.bindings.set(command, { command, keys });
-		if (this.config.debug) {
-			console.log(`[KeyboardManager] Binding set: ${command} -> ${keys.join(", ")}`);
-		}
-	}
-
-	setBindings(bindings: KeyBindings): void {
-		Object.entries(bindings).forEach(([command, keys]) => {
-			this.setBinding(command, keys);
-		});
-	}
-
-	removeBinding(command: string): void {
-		this.bindings.delete(command);
-	}
-
-	getBindings(): KeyBindings {
-		const bindings: KeyBindings = {};
-		for (const [command, binding] of this.bindings) {
-			bindings[command] = binding.keys;
-		}
-		return bindings;
 	}
 
 	loadPreset(preset: KeyboardPreset): void {
-		// 既存のバインディングをクリア
 		this.bindings.clear();
-
-		// プリセットのバインディングを設定
 		Object.entries(preset.bindings).forEach(([command, keys]) => {
 			this.setBinding(command, keys);
 		});
@@ -98,211 +57,184 @@ export class KeyboardManager implements IKeyboardManager {
 		}
 	}
 
+	getBindings(): KeyBindings {
+		const bindings: KeyBindings = {};
+		for (const [command, keys] of this.bindings) {
+			bindings[command] = keys;
+		}
+		return bindings;
+	}
+
+	// Context管理のデリゲート
 	pushContext(name: string, bindings?: KeyBindings): void {
-		const priority = this.contexts.size;
-		this.contexts.set(name, {
-			name,
-			bindings: bindings || undefined,
-			priority,
-		});
-		this.activeContext = name;
+		this.contextMixin.pushContext(name, bindings);
+		if (this.config.debug) {
+			console.log(`[KeyboardManager] Pushed context: ${name}`);
+		}
 	}
 
 	popContext(): void {
-		if (this.activeContext !== "default") {
-			this.contexts.delete(this.activeContext);
-			// 最も優先度の高いコンテキストを見つける
-			let maxPriority = -1;
-			let newContext = "default";
-
-			for (const [name, context] of this.contexts) {
-				if (context.priority > maxPriority) {
-					maxPriority = context.priority;
-					newContext = name;
-				}
-			}
-
-			this.activeContext = newContext;
-		}
-	}
-
-	registerCommand(name: string, handler: CommandHandler): void {
-		this.commandHandlers[name] = handler;
+		this.contextMixin.popContext();
 		if (this.config.debug) {
-			console.log(`[KeyboardManager] Command registered: ${name}`);
+			console.log("[KeyboardManager] Popped context");
 		}
 	}
 
-	unregisterCommand(name: string): void {
-		delete this.commandHandlers[name];
-	}
-
-	executeCommand(command: string, event: KeyboardEvent): boolean {
-		const handler = this.commandHandlers[command];
-		if (handler) {
-			const result = handler(event);
-			if (this.config.debug) {
-				console.log(`[KeyboardManager] Command executed: ${command} -> ${result}`);
-			}
-
-			// スクリーンリーダーにコマンド実行を通知
-			if (this.screenReader && result) {
-				const binding = this.bindings.get(command);
-				if (binding) {
-					const shortcutLabel = formatShortcutLabel(binding.keys);
-					this.screenReader.announce(`${command} (${shortcutLabel})`);
+	// KeyboardManager固有のメソッド
+	handleKeyDown(event: KeyboardEvent): boolean {
+		// キャッシュチェック
+		if (this.commandCache.has(event)) {
+			const cachedCommands = this.commandCache.get(event);
+			if (cachedCommands?.size) {
+				for (const command of cachedCommands) {
+					if (this.executeCommand(command, event)) {
+						return true;
+					}
 				}
 			}
-
-			return result;
-		}
-		return false;
-	}
-
-	handleKeyDown(event: KeyboardEvent): boolean {
-		// 入力フィールド内でのキーボード操作を無視する設定
-		if (!this.config.enableInInput && this.isInputElement(event.target)) {
 			return false;
 		}
 
 		const key = this.normalizeKey(event);
-		this.pressedKeys.add(key);
+		this.updateModifierState(key, true);
 
-		// スペースキーの特別処理
-		if (event.key === " " || event.key === "Space") {
+		// スペースキーのイベントを発火
+		if (key === "space") {
 			this.emit("space:down", event);
 		}
 
-		const command = this.findCommand(key);
-
-		if (command) {
-			event.preventDefault();
-			return this.executeCommand(command, event);
+		// コンテキスト内のバインディングを優先
+		const context = this.contextMixin.getCurrentContext();
+		if (context?.bindings) {
+			for (const [command, keys] of context.bindings) {
+				if (this.matchesBinding(event, keys as string[])) {
+					this.announceCommand(command);
+					return this.executeCommand(command, event);
+				}
+			}
 		}
 
+		// グローバルバインディングをチェック
+		const matchedCommands = new Set<string>();
+		for (const [command, binding] of this.bindings) {
+			// bindingはstring[]型
+			const keys = binding as string[];
+			if (this.matchesBinding(event, keys)) {
+				matchedCommands.add(command);
+				this.announceCommand(command);
+				if (this.executeCommand(command, event)) {
+					this.commandCache.set(event, matchedCommands);
+					return true;
+				}
+			}
+		}
+
+		this.commandCache.set(event, matchedCommands);
 		return false;
 	}
 
 	handleKeyUp(event: KeyboardEvent): boolean {
 		const key = this.normalizeKey(event);
-		this.pressedKeys.delete(key);
+		this.updateModifierState(key, false);
 
-		// スペースキーの特別処理
-		if (event.key === " " || event.key === "Space") {
+		// スペースキーのイベントを発火
+		if (key === "space") {
 			this.emit("space:up", event);
 		}
 
 		return false;
 	}
 
-	destroy(): void {
-		this.bindings.clear();
-		this.contexts.clear();
-		this.commandHandlers = {};
-		this.listeners.clear();
-		this.pressedKeys.clear();
-
-		// スクリーンリーダーのクリーンアップ
-		if (this.screenReader) {
-			this.screenReader.destroy();
-			this.screenReader = null;
-		}
+	isModifierActive(modifier: string): boolean {
+		return this.modifierState.get(modifier) || false;
 	}
 
 	// ヘルパーメソッド
 	private normalizeKey(event: KeyboardEvent): string {
-		const parts: string[] = [];
+		const key = event.key.toLowerCase();
+		return key === " " ? "space" : key;
+	}
 
-		// プラットフォーム非依存の修飾キー
-		if (event.ctrlKey || event.metaKey) parts.push("mod");
-		if (event.shiftKey) parts.push("shift");
-		if (event.altKey) parts.push("alt");
-
-		// キーを小文字に正規化
-		let key = event.key.toLowerCase();
-
-		// 特殊キーの正規化
-		const keyMap: Record<string, string> = {
-			" ": "space",
-			arrowup: "ArrowUp",
-			arrowdown: "ArrowDown",
-			arrowleft: "ArrowLeft",
-			arrowright: "ArrowRight",
-			enter: "Enter",
-			escape: "Escape",
-			delete: "Delete",
-			backspace: "Backspace",
-			tab: "Tab",
+	private updateModifierState(key: string, pressed: boolean): void {
+		const modifierMap: Record<string, string> = {
+			control: "ctrl",
+			meta: "cmd",
+			alt: "alt",
+			shift: "shift",
 		};
 
-		key = keyMap[key] || key;
-		parts.push(key);
-
-		return parts.join("+");
+		const modifier = modifierMap[key];
+		if (modifier) {
+			this.modifierState.set(modifier, pressed);
+		}
 	}
 
-	private findCommand(key: string): string | undefined {
-		// 現在のコンテキストのバインディングを優先
-		const context = this.contexts.get(this.activeContext);
-		if (context?.bindings) {
-			for (const [command, keys] of Object.entries(context.bindings)) {
-				if (keys.includes(key)) {
-					return command;
+	private matchesBinding(event: KeyboardEvent, keys: string[]): boolean {
+		const pressedKey = this.normalizeKey(event);
+		const modifiers = this.getModifiers(event);
+
+		for (const binding of keys) {
+			const parts = binding.toLowerCase().split("+");
+			const bindingKey = parts[parts.length - 1];
+			const bindingModifiers = parts.slice(0, -1);
+
+			if (bindingKey !== pressedKey) continue;
+
+			// 修飾キーのマッチング
+			const requiredMods = new Set(bindingModifiers);
+			const activeMods = new Set(modifiers);
+
+			// "mod"をプラットフォーム依存のキーに変換
+			if (requiredMods.has("mod")) {
+				requiredMods.delete("mod");
+				// metaKeyが押されている場合はcmd、ctrlKeyが押されている場合はctrlとして扱う
+				if (event.metaKey) {
+					requiredMods.add("cmd");
+				} else if (event.ctrlKey) {
+					requiredMods.add("ctrl");
 				}
 			}
-		}
 
-		// グローバルバインディングを検索
-		for (const [command, binding] of this.bindings) {
-			if (binding.keys.includes(key)) {
-				return command;
+			if (requiredMods.size !== activeMods.size) continue;
+
+			let allMatch = true;
+			for (const mod of requiredMods) {
+				if (!activeMods.has(mod)) {
+					allMatch = false;
+					break;
+				}
 			}
+
+			if (allMatch) return true;
 		}
 
-		return undefined;
+		return false;
 	}
 
-	private isInputElement(target: EventTarget | null): boolean {
-		if (!target || !(target instanceof HTMLElement)) return false;
-
-		const tagName = target.tagName.toLowerCase();
-		const editableElements = ["input", "textarea", "select"];
-
-		return (
-			editableElements.includes(tagName) ||
-			target.contentEditable === "true" ||
-			target.getAttribute("role") === "textbox"
-		);
+	private getModifiers(event: KeyboardEvent): string[] {
+		const modifiers: string[] = [];
+		if (event.ctrlKey) modifiers.push("ctrl");
+		if (event.metaKey) modifiers.push("cmd");
+		if (event.altKey) modifiers.push("alt");
+		if (event.shiftKey) modifiers.push("shift");
+		return modifiers;
 	}
 
-	// イベントエミッター機能
-	on(event: string, listener: (data: any) => void): void {
-		if (!this.listeners.has(event)) {
-			this.listeners.set(event, new Set());
+	private announceCommand(command: string): void {
+		if (this.announcer) {
+			const message = `${command.replace(/([A-Z])/g, " $1").toLowerCase()} activated`;
+			this.announcer.announce(message);
 		}
-		this.listeners.get(event)?.add(listener);
 	}
 
-	off(event: string, listener: (data: any) => void): void {
-		this.listeners.get(event)?.delete(listener);
-	}
-
-	private emit(event: string, data: any): void {
-		this.listeners.get(event)?.forEach((listener) => {
-			listener(data);
-		});
-	}
-
-	// 公開メソッド
-	isKeyPressed(key: string): boolean {
-		return this.pressedKeys.has(key);
-	}
-
-	getActiveBindings(): KeyBinding[] {
-		return Array.from(this.bindings.values());
-	}
-
-	getActiveContext(): string {
-		return this.activeContext;
+	override destroy(): void {
+		super.destroy();
+		this.contextMixin.clearContexts();
+		this.modifierState.clear();
+		this.commandCache = new WeakMap();
+		this.announcer?.destroy();
 	}
 }
+
+// デフォルトエクスポート
+export default KeyboardManager;
